Compute conversion scale in datetime bindings

+++ python/numpy.h
@@ -205,4 +205,4 @@ void copy_array_into_view(const std::vector<SourceDType> &src,
   std::copy(begin(src), end(src), dst.begin());
 }
 
-core::time_point make_time_point(const py::buffer &buffer, int64_t scale = 1);
+core::time_point make_time_point(const py::buffer &buffer, double scale = 1.0);
